Losing the will to live...

+++ Tasks.Bounce/db_create_development.cs
@@ -1,52 +1,29 @@
 ﻿using System;
 using System.IO;
 using System.Linq;
+using System.Text.RegularExpressions;
 using Bounce.Framework;
-using Codeifier.OrangeCMS.Domain.Providers;
-using Codeifier.OrangeCMS.Repositories;
-using Fixtures;
-using OrangeCMS.Application.Providers;
-using OrangeCMS.Application.Services;
-using OrangeCMS.Domain;
+using CodeKinden.OrangeCMS.Application.Providers;
+using CodeKinden.OrangeCMS.Domain.Models;
+using CodeKinden.OrangeCMS.Domain.Services;
+using CodeKinden.OrangeCMS.Fixtures;
+using CodeKinden.OrangeCMS.Repositories;
 
 namespace CodeKinden.OrangeCMS.Tasks.Bounce
 {
     public class db_create_development
     {
         [Task(Command = "db:create:development", Description = "Creates the development database.")]
-        public void Execute(string sourceConnectionString, string targetConnectionString, string boundaries = "", string customers = "", string customersOrange = "")
+        public void Execute(string targetConnectionString, int maxBoundaries = int.MaxValue, int maxCustomers = int.MaxValue)
         {
             DB.CreateOrReplaceDatabase(targetConnectionString);
 
             DB.RunDatabaseMigrations(targetConnectionString);
-
-            var boundaryData = String.IsNullOrEmpty(boundaries) ? FixtureManager.Extract("CodeKinden.OrangeCMS.Fixtures.Countries.zip") : boundaries;
-
-            Console.WriteLine("Using the boundary data contained in " + boundaryData);
-
-            if (!File.Exists(boundaryData))
-            {
-                throw new Exception(String.Format("The file '{0}' could not be found.", boundaryData));
-            }
-
-            var customerData = customers == "" ? FixtureManager.Extract("CodeKinden.OrangeCMS.Fixtures.Customers.csv") : customers;
-
-            if (!File.Exists(customerData))
-            {
-                throw new Exception(String.Format("The file '{0}' could not be found.", customerData));
-            }
-
-            var customerOrangeData = customers == "" ? FixtureManager.Extract("CodeKinden.OrangeCMS.Fixtures.Customers.FromOrange.csv") : customersOrange;
-
-            if (!File.Exists(customerOrangeData))
-            {
-                throw new Exception(String.Format("The file '{0}' could not be found.", customerOrangeData));
-            }
-
-            GenerateData(targetConnectionString, boundaryData, customerData, customerOrangeData);
+            
+            GenerateData(targetConnectionString, maxBoundaries, maxCustomers);
         }
         
-        private static void GenerateData(string connectionString, string boundaryData, string customerData, string customerOrangeData)
+        private static void GenerateData(string connectionString, int maxBoundaries, int maxCustomers)
         {
             var dbContextScope = new DbContextScope(connectionString);
 
@@ -64,21 +41,54 @@ namespace CodeKinden.OrangeCMS.Tasks.Bounce
 
             dbContext.SaveChanges();
 
-            var boundaryService = new BoundaryService(dbContextScope);
-            boundaryService.SaveBoundariesInZip("name", boundaryData);
-            Console.WriteLine("The boundaries have been saved to the database.");
+            SaveBoundaryData(maxBoundaries, dbContextScope, "CodeKinden.OrangeCMS.Fixtures.Quartiers.Abidjan.Commune.zip", "quartier_8");
 
+            SaveBoundaryData(maxBoundaries, dbContextScope, "CodeKinden.OrangeCMS.Fixtures.Quartiers_092010.zip", "Descriptio", x => Regex.Match(x, (@"<br/>Quartier : .(.+)<br/>")).Groups[1].Value);
+
+            var customerData = FixtureManager.Extract("CodeKinden.OrangeCMS.Fixtures.Customers.csv");
+
+            if (!File.Exists(customerData))
+            {
+                throw new Exception(String.Format("The file '{0}' could not be found.", customerData));
+            }
+
+            var customerOrangeData = FixtureManager.Extract("CodeKinden.OrangeCMS.Fixtures.Customers.FromOrange.csv");
+
+            if (!File.Exists(customerOrangeData))
+            {
+                throw new Exception(String.Format("The file '{0}' could not be found.", customerOrangeData));
+            }
+            
             dbContext.Database.ExecuteSqlCommand("DELETE FROM dbo.Boundaries WHERE Shape.STIsValid() = 0");
 
-            SaveCustomerData(dbContextScope, customerData);
-            SaveCustomerData(dbContextScope, customerOrangeData);
+            SaveCustomerData(dbContextScope, customerData, maxCustomers);
+            SaveCustomerData(dbContextScope, customerOrangeData, maxCustomers);
             Console.WriteLine("The customers have been saved to the database.");
         }
-        
-        private static void SaveCustomerData(IDbContextScope dbContextScope, string customerData)
+
+        private static void SaveBoundaryData(int maxBoundaries, IDbContextScope dbContextScope, string resourceName, string columnName, Func<string, string> columnNameParser)
+        {
+            var boundaryData = FixtureManager.Extract(resourceName);
+
+            if (!File.Exists(boundaryData))
+            {
+                throw new Exception(String.Format("The file '{0}' could not be found.", boundaryData));
+            }
+
+            var boundaryService = new BoundaryService(dbContextScope);
+            boundaryService.SaveBoundariesInZip(columnName, columnNameParser, boundaryData, maxBoundaries);
+            Console.WriteLine("The boundaries from '{0}' have been saved to the database.", resourceName);
+        }
+
+        private static void SaveBoundaryData(int maxBoundaries, IDbContextScope dbContextScope, string resourceName, string columnName)
+        {
+            SaveBoundaryData(maxBoundaries, dbContextScope, resourceName, columnName, x => x);
+        }
+
+        private static void SaveCustomerData(IDbContextScope dbContextScope, string customerData, int maxCustomers)
         {
             var customerService = new CustomerService(dbContextScope);
-            var customers = customerService.Import(customerData);
+            var customers = customerService.Import(customerData, maxCustomers);
             customerService.Save(customers.ToArray());
         }
     }
